Increment basket from partner screen

+++ app/src/main/java/ooo/cron/delivery/data/network/models/PartnerProductsRes.kt
@@ -9,7 +9,7 @@ package ooo.cron.delivery.data.network.models
 data class PartnerProductsRes(
     var id: String,
     var name: String,
-    var cost: Int,
+    var cost: Double,
     var photo: String,
     var portionSize: String,
     var description: String,
